code/LlamaIndex/requirements.txt
llama-index
python-dotenv
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-llms-mistralai
llama-index-vector-stores-chroma